Updated some more code and added content to the OM paper. Need to finish joint stuff, then collect data and put it in the report, and then write analysis.

+++ OptMethods/project/code/Avalanche/wt.m
@@ -4,7 +4,7 @@ function w = wt(x, n)
         w = 0;
     else
         for i = 1:n
-            w = w + bitget(uint8(x), i);	
+            w = w + bitget(x, i);	
         end
     end
 end
