+++ OptMethods/project/code/Joint/order4/wt.m
@@ -1,6 +1,6 @@
 function w = wt(x, n)
 	w = 0;
 	for i = 1:n
-		w = w + bitget(uint8(x), i);	
+		w = w + bitget(x, i);	
     end
 end
